Add stats and charts to variable yield details
The cards already existed on Analytics, so they move to components/
and both pages render the same one. Each takes only a peggedToken;
the page supplies the container.
