+++ web/src/components/icons/clockRevertedIcon.tsx
@@ -1,0 +1,21 @@
+import type { ComponentProps } from "react";
+
+export const ClockRevertedIcon = (props: ComponentProps<"svg">) => (
+  <svg
+    fill="none"
+    height="16"
+    viewBox="0 0 16 16"
+    width="16"
+    xmlns="http://www.w3.org/2000/svg"
+    {...props}
+  >
+    <path
+      d="M8.11602 3.58789C8.30077 3.58789 8.47795 3.66128 8.60859 3.79192C8.73923 3.92255 8.81262 4.09973 8.81262 4.28448V7.71171L10.4659 9.36496C10.5343 9.42873 10.5892 9.50563 10.6273 9.59108C10.6653 9.67653 10.6858 9.76877 10.6875 9.8623C10.6891 9.95583 10.6719 10.0487 10.6369 10.1355C10.6018 10.2222 10.5497 10.301 10.4835 10.3672C10.4174 10.4333 10.3386 10.4854 10.2519 10.5205C10.1651 10.5555 10.0722 10.5727 9.97869 10.5711C9.88516 10.5694 9.79292 10.549 9.70747 10.5109C9.62202 10.4728 9.54512 10.4179 9.48134 10.3495L7.62377 8.49189C7.49309 8.36138 7.4196 8.18432 7.41943 7.99964V4.28448C7.41943 4.09973 7.49282 3.92255 7.62346 3.79192C7.7541 3.66128 7.93128 3.58789 8.11602 3.58789Z"
+      fill="currentColor"
+    />
+    <path
+      d="M11.4711 11.8476C10.5218 12.6743 9.30178 13.1232 8.04309 13.109C6.7844 13.0948 5.57484 12.6184 4.64447 11.7705C3.71639 10.9245 3.13247 9.76622 3.00423 8.51695C2.99622 8.42493 2.96998 8.33541 2.92704 8.25362C2.88411 8.17183 2.82533 8.0994 2.75414 8.04054C2.68295 7.98168 2.60076 7.93757 2.51235 7.91077C2.42395 7.88397 2.3311 7.87502 2.23921 7.88444C2.14732 7.89387 2.05821 7.92147 1.97709 7.96566C1.89597 8.00984 1.82444 8.06972 1.76668 8.1418C1.70891 8.21389 1.66606 8.29674 1.64062 8.38554C1.61518 8.47435 1.60765 8.56732 1.61848 8.65906C1.78186 10.249 2.52515 11.723 3.7064 12.7996C4.8891 13.8775 6.42667 14.4832 8.02678 14.5015C9.62688 14.5198 11.1779 13.9495 12.385 12.899C13.5911 11.8502 14.3679 10.3933 14.5667 8.80743C14.7655 7.22152 14.3723 5.61801 13.4624 4.30397C12.864 3.44339 12.0673 2.73951 11.1396 2.25182C10.2118 1.76412 9.18024 1.50694 8.1321 1.50203C7.08397 1.49711 6.05006 1.74459 5.11775 2.22355C4.18544 2.70252 3.38215 3.39889 2.77575 4.25382V2.76125C2.77575 2.5765 2.70236 2.39932 2.57172 2.26869C2.44109 2.13805 2.26391 2.06466 2.07916 2.06466C1.89441 2.06466 1.71723 2.13805 1.5866 2.26869C1.45596 2.39932 1.38257 2.5765 1.38257 2.76125V6.0083C1.38257 6.39282 1.69464 6.70489 2.07916 6.70489H5.32992C5.51467 6.70489 5.69185 6.6315 5.82248 6.50086C5.95312 6.37022 6.02651 6.19304 6.02651 6.0083C6.02651 5.82355 5.95312 5.64637 5.82248 5.51573C5.69185 5.3851 5.51467 5.31171 5.32992 5.31171H3.74448C4.21907 4.54658 4.88746 3.92049 5.68193 3.49685C6.79197 2.90463 8.08077 2.74044 9.30383 3.03542C10.5269 3.3304 11.5992 4.06404 12.3172 5.09716C13.0323 6.12909 13.3413 7.38868 13.185 8.63441C13.0287 9.88014 12.418 11.0243 11.4701 11.8476"
+      fill="currentColor"
+    />
+  </svg>
+);
