+++ web/src/components/icons/trendingUpIcon.tsx
@@ -1,9 +1,18 @@
-export const TrendingUpIcon = () => (
-  <svg fill="none" height="16" width="16" xmlns="http://www.w3.org/2000/svg">
+import type { ComponentProps } from "react";
+
+export const TrendingUpIcon = (props: ComponentProps<"svg">) => (
+  <svg
+    fill="none"
+    height="16"
+    viewBox="0 0 16 16"
+    width="16"
+    xmlns="http://www.w3.org/2000/svg"
+    {...props}
+  >
     <path
       clipRule="evenodd"
       d="M9.808 4.057a.75.75 0 0 1 .92-.527l3.116.85a.75.75 0 0 1 .528.914l-.823 3.121a.75.75 0 0 1-1.45-.382l.337-1.28a23.487 23.487 0 0 0-3.609 3.055.75.75 0 0 1-1.07.01L6 8.06l-3.72 3.72a.75.75 0 1 1-1.06-1.06l4.25-4.251a.75.75 0 0 1 1.06 0l1.756 1.755a25.017 25.017 0 0 1 3.508-2.85l-1.46-.398a.75.75 0 0 1-.526-.919Z"
-      fill="#416BFF"
+      fill="currentColor"
       fillRule="evenodd"
     />
   </svg>
