+++ web/src/components/yieldCard.tsx
@@ -1,21 +1,24 @@
+import { AllocationCard } from "components/allocationCard";
+import { PieChartIcon } from "components/icons/pieChartIcon";
 import { useAnalyticsTreasury } from "hooks/useAnalyticsTreasury";
+import { usePeggedTokensByGateway } from "hooks/usePeggedTokensByGateway";
 import { usePrices } from "hooks/usePrices";
 import { useWhitelistedTokensByGateway } from "hooks/useWhitelistedTokensByGateway";
 import { useTranslation } from "react-i18next";
 import type { TokenWithGateway } from "types";
-
-import { PieChartIcon } from "../icons/pieChartIcon";
-import { assignColor, toReserveBufferAmount, toYieldItems } from "../utils";
-
-import { AllocationCard } from "./allocationCard";
+import {
+  assignColor,
+  toReserveBufferAmount,
+  toYieldItems,
+} from "utils/allocations";
 
 type Props = {
   peggedToken: TokenWithGateway | undefined;
-  peggedTokenError: boolean;
 };
 
-export const YieldCard = function ({ peggedToken, peggedTokenError }: Props) {
+export const YieldCard = function ({ peggedToken }: Props) {
   const { t } = useTranslation();
+  const { isError: isPeggedTokensError } = usePeggedTokensByGateway();
   const { data: whitelistedTokens, isError: isWhitelistedTokensError } =
     useWhitelistedTokensByGateway(peggedToken?.gatewayAddress);
   const {
@@ -26,7 +29,7 @@ export const YieldCard = function ({ peggedToken, peggedTokenError }: Props) {
   const { data: prices, isError: isPricesError } = usePrices();
 
   const isError = [
-    peggedTokenError,
+    isPeggedTokensError,
     isWhitelistedTokensError,
     isTreasuryError,
     isPricesError,
@@ -57,23 +60,23 @@ export const YieldCard = function ({ peggedToken, peggedTokenError }: Props) {
           {
             amount: bufferAmount,
             color: assignColor(yieldItems.length),
-            label: t("pages.analytics.reserve-buffer-label"),
+            label: t("common.charts.reserve-buffer-label"),
           },
         ]
       : yieldItems;
 
   const value =
     yieldItems && treasury
-      ? t("pages.analytics.yield-value", { count: yieldItems.length })
+      ? t("common.charts.yield-value", { count: yieldItems.length })
       : "";
 
   return (
     <AllocationCard
-      icon={<PieChartIcon />}
+      icon={<PieChartIcon className="text-blue-500" />}
       isError={isError}
       isLoading={isLoading}
       items={items}
-      label={t("pages.analytics.yield-label")}
+      label={t("common.charts.yield-label")}
       value={value}
     />
   );
