+++ web/src/hooks/useDeposit.ts
@@ -6,8 +6,7 @@ import { useMutation, useQueryClient } from "@tanstack/react-query";
 import type { DepositEvents } from "@vetro-protocol/gateway";
 import { deposit } from "@vetro-protocol/gateway/actions";
 import type { EventEmitter } from "events";
-import type { TreasuryToken } from "pages/analytics/types";
-import type { TokenWithGateway } from "types";
+import type { TokenWithGateway, TreasuryToken } from "types";
 import { getReceivedAmount } from "utils/receipt";
 import { applySlippage } from "utils/slippage";
 import { type Address, isAddressEqual } from "viem";
