+++ web/src/i18n/locales/es/translation.json
@@ -5,11 +5,17 @@
     "approve-10x": "Aprobar 10x",
     "approve-10x-tooltip": "Apruebe hasta 10× el monto de su transacción, reduciendo futuros pasos de aprobación. Puede revocarlo en cualquier momento.",
     "charts": {
+      "apy-label": "TEA",
       "period-1-month": "1 mes",
       "period-1-week": "1 semana",
       "period-1-year": "1 año",
       "period-3-month": "3 meses",
-      "reload-chart": "Recargar gráfico"
+      "reload-chart": "Recargar gráfico",
+      "reserve-buffer-label": "Reserva de liquidez",
+      "share-ratio-label": "Ratio {{shareSymbol}} / {{peggedSymbol}}",
+      "yield-label": "Asignación de yield",
+      "yield-value": "{{count}} estrategia",
+      "yield-value_other": "{{count}} estrategias"
     },
     "clear-search": "Limpiar búsqueda",
     "close": "Cerrar",
@@ -56,20 +62,12 @@
   },
   "pages": {
     "analytics": {
-      "apy-label": "TEA",
       "collateralization-ratio-label": "Ratio de colateralización",
-      "exit-queue-label": "Total {{symbol}} en período de espera de retiro",
       "liquid-reserves-label": "Reservas líquidas",
-      "reserve-buffer-label": "Reserva de liquidez",
-      "share-ratio-label": "Ratio {{shareSymbol}} / {{peggedSymbol}}",
-      "staked-label": "Total {{symbol}} depositado",
       "strategic-reserves-label": "Reservas estratégicas",
       "surplus-label": "Excedente",
       "title": "Monitorear analíticas del protocolo",
-      "tvl-label": "Valor Total Bloqueado (TVL)",
-      "yield-label": "Asignación de yield",
-      "yield-value": "{{count}} estrategia",
-      "yield-value_other": "{{count}} estrategias"
+      "tvl-label": "Valor Total Bloqueado (TVL)"
     },
     "borrow": {
       "activity": {
@@ -273,6 +271,7 @@
         "instant-withdraw-text": "{{amount}} {{symbol}} retirado",
         "request-withdraw-text": "{{amount}} {{symbol}} enviado al exit vault"
       },
+      "exit-cooldown-label": "{{symbol}} en período de espera de salida",
       "exit-tickets": {
         "actions": "Acciones",
         "col-cooldown-left": "Tiempo restante",
@@ -387,6 +386,7 @@
         "you-will-stake": "Usted va a hacer stake de",
         "you-will-withdraw": "Usted va a retirar"
       },
+      "staked-label": "{{symbol}} staked",
       "stats": {
         "earned-amount": "Su monto ganado",
         "from-pools": "De {{count}} fondo",
@@ -395,6 +395,7 @@
       },
       "title": "Haga stake de activos para ganar un rendimiento variable.",
       "variable-yield": {
+        "apy": "TEA",
         "header-title": "Rendimiento variable",
         "title": "Fondos de rendimiento variable"
       }
