--- web/src/pages/analytics/icons/exitQueueIcon.tsx
@@ -1,8 +1,0 @@
-export const ExitQueueIcon = () => (
-  <svg fill="none" height="16" width="16" xmlns="http://www.w3.org/2000/svg">
-    <path
-      d="M8.75 6h-1.5V3.56L6.03 4.78a.75.75 0 0 1-1.06-1.06l2.5-2.5a.75.75 0 0 1 1.06 0l2.5 2.5a.75.75 0 1 1-1.06 1.06L8.75 3.56V6H11a2 2 0 0 1 2 2v5a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V8a2 2 0 0 1 2-2h2.25v5.25a.75.75 0 1 0 1.5 0V6Z"
-      fill="#416BFF"
-    />
-  </svg>
-);
