--- web/src/pages/analytics/types.ts
@@ -1,20 +1,0 @@
-import type { Address } from "viem";
-
-// `amount` is used for bar proportions (via CSS flex) and formatted by AllocationLegend.
-export type AllocationItem = {
-  amount: number;
-  color: string;
-  label: string;
-  logoURI?: string;
-  tooltip?: string;
-};
-
-// Raw response shape from GET /analytics/treasury.
-export type TreasuryToken = {
-  activeStrategies: { name: string; totalDebt: string }[];
-  latestPrice: string;
-  priceDecimals: number;
-  tokenAddress: Address;
-  totalDebt: string;
-  withdrawable: string;
-};
