+++ web/src/pages/earn/components/earnedAmountStat/index.tsx
@@ -1,9 +1,8 @@
 import { InfoCard } from "components/base/infoCard";
+import { TrendingUpIcon } from "components/icons/trendingUpIcon";
 import { useEarnedAmountUsd } from "hooks/useEarnedAmountUsd";
 import { useTranslation } from "react-i18next";
 import { formatUsd } from "utils/currency";
-
-import { TrendingUpIcon } from "../../icons/trendingUpIcon";
 
 export function EarnedAmountStat() {
   const { t } = useTranslation();
@@ -11,7 +10,7 @@ export function EarnedAmountStat() {
   return (
     <InfoCard
       {...useEarnedAmountUsd()}
-      icon={<TrendingUpIcon />}
+      icon={<TrendingUpIcon className="text-blue-500" />}
       label={t("pages.earn.stats.earned-amount")}
       render={formatUsd}
     />
