--- web/src/pages/earn/icons/boltIcon.tsx
@@ -1,10 +1,0 @@
-export const BoltIcon = () => (
-  <svg fill="none" height="16" width="16" xmlns="http://www.w3.org/2000/svg">
-    <path
-      clipRule="evenodd"
-      d="M9.58 1.077a.75.75 0 0 1 .405.82L9.165 6h4.085a.75.75 0 0 1 .567 1.241l-6.5 7.5a.75.75 0 0 1-1.302-.638L6.835 10H2.75a.75.75 0 0 1-.567-1.24l6.5-7.5a.75.75 0 0 1 .897-.183Z"
-      fill="#416BFF"
-      fillRule="evenodd"
-    />
-  </svg>
-);
