+++ web/src/types/index.ts
@@ -9,3 +9,13 @@ export type BridgeableToken = Token & {
   oftAdapterAddress?: Address;
   sharedDecimals: number;
 };
+
+// Raw response shape from GET /analytics/treasury.
+export type TreasuryToken = {
+  activeStrategies: { name: string; totalDebt: string }[];
+  latestPrice: string;
+  priceDecimals: number;
+  tokenAddress: Address;
+  totalDebt: string;
+  withdrawable: string;
+};
