+++ web/stories/allocationLegend.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
-import { AllocationLegend } from "../src/pages/analytics/components/allocationCard/allocationLegend";
+import { AllocationLegend } from "../src/components/allocationCard/allocationLegend";
 import { formatUsd } from "../src/utils/currency";
 
 const meta = {
